Fix shellcheck finding SC2086 in .gitlab-ci.yml
2nd attempt. Double-quotes passes all files as a single string instead
as separate parameters to shellcheck. $files was actually an array
constructed through globbing. It actually missed some files because
they were an extra level deep.

Replaced with a [anti-fragile] find command to creates the array.
The array expansion in the shellcheck command can then be double-quoted.

Also needed to adjust the check
so that the 0th element was checked explicitly to avoid another
warning.